chimera: avoid extra db lookups on create and remove
Acked-by: Karsten Schwank
Acked-by: Gerd Behrmann
Target: master
Require-book: no
Require-notes: no

+++ modules/chimera/src/main/java/org/dcache/chimera/FsSqlDriver.java
@@ -257,6 +257,9 @@ class FsSqlDriver {
     void remove(Connection dbConnection, FsInode parent, String name) throws ChimeraFsException, SQLException {
 
         FsInode inode = inodeOf(dbConnection, parent, name);
+        if (inode == null || inode.type() != FsInodeType.INODE) {
+            throw new FileNotFoundHimeraFsException("Not a file.");
+        }
 
         if (inode.isDirectory()) {
             removeDir(dbConnection, parent, inode, name);
